pyproject.toml
[project]
name = "multi-region-dashboard"
version = "1.0.0"
description = "Interactive dashboard for testing PostgreSQL connectivity across multiple Aiven regions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "asyncpg>=0.29.0",
  "jinja2>=3.1.2",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "cryptography>=46.0.3",
  "starlette>=0.50.0",
  "annotated-types>=0.7.0",
  "bcrypt>=5.0.0",
  "python-multipart",
]

[project.optional-dependencies]
dev = ["black>=23.0.0", "ruff>=0.1.0", "mypy>=1.7.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP", # pyupgrade
]
ignore = [
  "E501", # line too long, handled by black
]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["asyncpg.*"]
ignore_missing_imports = true

[dependency-groups]
dev = ["black>=25.11.0", "mypy>=1.19.0", "ruff>=0.14.8"]
